Bank Program C++
Banks have many different types of accounts often with different rules
for fees
associated with transactions such as withdrawals. Customers are allowed
to
transfer funds between accounts incurring the appropriate fees
associated with
withdrawals of funds from one account. (You may create a customer data
which
maybe stored in a binary or text file)
Write a program with a base class for a bank account and one derived
class (as
described below) representing an account with a different rule for
withdrawing
funds. Also write a function that transfers funds from one account of
any type to
another. A transfer is a withdrawal from one account and a deposit into
the other.
Since the transfer can be done at any time with any type of account the
withdraw
function in the classes must be virtual. Write a main program that
creates three
accounts one from each class and tests the transfer function.
For the classes, create a base class called BankAccount that has the
name (a
string) and number (an int) of the owner of the account and the balance
in the
account (double) as data members. Include member functions deposit and
withdraw (each with a double for the amount as an argument) and assessor
functions getName and getBalance. Deposit will add the amount to the
balance (assuming the amount is nonnegative) and withdraw will subtract
the
amount from the balance (assuming the amount is nonnegative and less
than or
equal to the balance). Also create a class called MoneyMarketAccount
that is
derived from BankAccount. In a MoneyMarketAccount the user gets 2 free
withdrawals in a given period of time (don’t worry about the time for
this
problem). After the free withdrawals have been used, a withdrawal fee of
R2.00
is deducted from the balance per withdrawal. Hence, the class must have
a data
member to keep track of the number of withdrawals. It also must override
the
withdraw definition. For all two classes, the withdraw function should
return an
integer indicating the status (either ok or insufficient funds for the
withdrawal to
take place). For the purpose of this exercise do not worry about when
and how
interest is paid